djangoRest/mysite2/requirements.txt
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-twilio==0.13.0.post1
djangorestframework==3.12.1
